# dnsname: Optimize parsing of uncompressed labels