@rgacogne 2024-01-26
I just pushed a commit applying your suggestions, Otto, thanks!